Under Windows sizeof(int)==4 on both x86 and x64, so switch from int to size_t with checks for overflow. Add some overflow checks before calling malloc().

+++ 0.11/lib/backupclient/BackupStoreFilename.h
@@ -101,7 +101,7 @@ protected:
 #define BACKUPSTOREFILENAME_GET_SIZE(hdr)		(( ((uint8_t)((hdr)[0])) | ( ((uint8_t)((hdr)[1])) << 8)) >> 2)
 #define BACKUPSTOREFILENAME_GET_ENCODING(hdr)	(((hdr)[0]) & 0x3)
 
-#define BACKUPSTOREFILENAME_MAKE_HDR(hdr, size, encoding)		{uint16_t h = (((uint16_t)size) << 2) | (encoding); ((hdr)[0]) = h & 0xff; ((hdr)[1]) = h >> 8;}
+#define BACKUPSTOREFILENAME_MAKE_HDR(hdr, size, encoding)		{uint16_t h = (((uint16_t)size) << 2) | (uint16_t)(encoding); ((hdr)[0]) = h & 0xff; ((hdr)[1]) = h >> 8;}
 
 #endif // BACKUPSTOREFILENAME__H
