+++ 0.11/lib/common/ReadGatherStream.h
@@ -32,12 +32,12 @@ private:
 	ReadGatherStream &operator=(const ReadGatherStream &);
 public:
 
-	int AddComponent(IOStream *pStream);
-	void AddBlock(int Component, pos_type Length, bool Seek = false, pos_type SeekTo = 0);
+	size_t AddComponent(IOStream *pStream);
+	void AddBlock(size_t Component, pos_type Length, bool Seek = false, pos_type SeekTo = 0);
 
-	virtual int Read(void *pBuffer, int NBytes, int Timeout = IOStream::TimeOutInfinite);
+	virtual size_t Read(void *pBuffer, size_t NBytes, int Timeout = IOStream::TimeOutInfinite);
 	virtual pos_type BytesLeftToRead();
-	virtual void Write(const void *pBuffer, int NBytes);
+	virtual void Write(const void *pBuffer, size_t NBytes);
 	virtual bool StreamDataLeft();
 	virtual bool StreamClosed();
 	virtual pos_type GetPosition() const;
@@ -50,7 +50,7 @@ private:
 	{
 		pos_type mLength;
 		pos_type mSeekTo;
-		int mComponent;
+		size_t mComponent;
 		bool mSeek;
 	} Block;
