+++ 0.11/lib/common/SelfFlushingStream.h
@@ -41,7 +41,7 @@ private:
 	SelfFlushingStream(const IOStream& rToCopy);
 	
 public:
-	virtual int Read(void *pBuffer, int NBytes,
+	virtual size_t Read(void *pBuffer, size_t NBytes,
 		int Timeout = IOStream::TimeOutInfinite)
 	{
 		return mrSource.Read(pBuffer, NBytes, Timeout);
@@ -50,7 +50,7 @@ public:
 	{
 		return mrSource.BytesLeftToRead();
 	}
-	virtual void Write(const void *pBuffer, int NBytes)
+	virtual void Write(const void *pBuffer, size_t NBytes)
 	{
 		mrSource.Write(pBuffer, NBytes);
 	}
